pyproject.toml
[tool.poetry]
name = "online-tech-store"
version = "0.1.0"
description = ""
authors = ["Bazavod <eugeny.bazavod@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.6"
pandas = "^2.3.2"
requests = "^2.32.5"
openpyxl = "^3.1.5"
python-dotenv = "^1.1.1"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.3.0"
ipython = "^9.5.0"
redis = "^6.4.0"
djangorestframework-simplejwt = "^5.5.1"
django-filter = "^25.1"
coverage = "^7.10.6"
gunicorn = "^23.0.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

[tool.black]
line-length = 160
exclude = '''(\.git)'''

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
